requirements.txt
Flask==1.1.1
pycryptodomex==3.9.9
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
